src/rules_maker/extractors/ml_extractor.py
"""
Placeholder for ML content extractor.
"""

from typing import Dict, List, Any
from bs4 import BeautifulSoup
from .base import ContentExtractor
from ..models import ContentSection


class MLContentExtractor(ContentExtractor):
    """ML-based content extraction using trained models."""
    
    def extract(self, soup: BeautifulSoup, url: str) -> Dict[str, Any]:
        """Extract structured content (placeholder)."""
        raise NotImplementedError("ML extractor not yet implemented")
    
    def extract_sections(self, soup: BeautifulSoup, url: str) -> List[ContentSection]:
        """Extract content sections (placeholder)."""
        raise NotImplementedError("ML extractor not yet implemented")
